World map Coloring using CSP in AI
Map coloring is a classic problem in the field of Artificial Intelligence (AI) and falls under the category of Constraint Satisfaction Problems (CSPs). The objective of this problem is to assign colors to different regions on a map in such a way that no two adjacent regions share the same color.The goal of the project is to advance the field of 
geographic information systems by using artificial intelligence to solve combinatorial optimization challenges. The abstract approach to solving the Map Colouring Problem in CSP involves defining a set of variables representing the regions, each with a domain of possible colours. Following that, constraints are provided to ensure that neighbouring regions have different colouring.